move all state structs into RefreshTypes
