--- src/Graphics/Structs/BlendConstants.cs
@@ -1,10 +1,0 @@
-﻿namespace MoonWorks.Graphics
-{
-	public struct BlendConstants
-	{
-		public float R;
-		public float G;
-		public float B;
-		public float A;
-	}
-}
